Working Directory/non-interactive -Credential sup
Add WorkingDirectory to the `ProcessInfo` outputs used in this module.
This can be used to get the current working directory of another
process, even ones that are not your own.

Fixes `Start-ProcessWith ... -Credential $cred` support when running in
a non-interactive process by explicitly inheriting the current
station/desktop rather than having Windows try to create one but fail.

Updates the build scripts to conform to newer standards and fixes.