Added Invoke-Process* cmdlets (#10)
Adds the Invoke-ProcessEx and Invoke-ProcessWith cmdlets which can be
used to invoke a process and capture the output inside PowerShell. This
is similar to invoking an executable normally in PowerShell but with the
extra functionality exposed by the underlying CreateProcess* APIs in
Windows.

Also adds -ArgumentEscaping which can be used to escape arguments in the
MSI property format.